compose: deprecate old compose implementation and provide options for new implementation
